% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDropCorr.R
\name{mlr_pipeops_dropcor}
\alias{mlr_pipeops_dropcor}
\alias{PipeOpDropCorr}
\title{Drop correlated features}
\description{
This PipeOp removes features that are highly correlated with other features.
}
\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{use} :: \code{character(1)}\cr
A character string indicating how to handle missing values. Possible values are
\code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, and \code{"pairwise.complete.obs"}.
Default is \code{"everything"}.
\item \code{method} :: \code{character(1)}\cr
A character string indicating the method to compute the correlation. Possible values are
\code{"pearson"}, \code{"kendall"}, and \code{"spearman"}. Default is \code{"pearson"}.
\item \code{cutoff} :: \code{numeric(1)}\cr
A numeric value indicating the correlation threshold above which features are considered
highly correlated. Default is \code{0.99}.
}
}

\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpDropCorr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpDropCorr-new}{\code{PipeOpDropCorr$new()}}
\item \href{#method-PipeOpDropCorr-clone}{\code{PipeOpDropCorr$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpDropCorr-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpDropCorr-new}{}}}
\subsection{Method \code{new()}}{
Constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpDropCorr$new(id = "dropcor", param_vals = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"drop.nacol"}.}

\item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction. Default \code{list()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpDropCorr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpDropCorr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpDropCorr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
