% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpUniform.R
\name{mlr_pipeops_uniform}
\alias{mlr_pipeops_uniform}
\alias{PipeOpUniform}
\title{Uniformize Numeric Features via ECDF}
\description{
This PipeOp transforms numeric features by applying the empirical cumulative distribution
function (ECDF) to each predictor. This maps the feature values to the [0, 1] interval.
If a grouping variable is defined in the task (via the group role), uniformization is applied
separately within each group.
}
\section{Parameters}{

The parameters are inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameter:
\itemize{
\item \code{na.rm} :: \code{logical(1)}\cr
Whether to ignore missing values when computing the ECDF. Default is \code{TRUE}.
}
}

\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpUniform}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpUniform-new}{\code{PipeOpUniform$new()}}
\item \href{#method-PipeOpUniform-clone}{\code{PipeOpUniform$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpUniform-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpUniform-new}{}}}
\subsection{Method \code{new()}}{
Constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpUniform$new(id = "uniformization", param_vals = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"uniformization"}.}

\item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings. Default is \code{list()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpUniform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpUniform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpUniform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
