% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpWinsorize.R
\name{mlr_pipeops_winsorize}
\alias{mlr_pipeops_winsorize}
\alias{PipeOpWinsorize}
\title{Winsorize Numeric Features}
\description{
This PipeOp performs winsorization on numeric features by capping values below the lower
percentile and above the upper percentile. If a grouping variable is defined in the task
(via the group role), the winsorization is applied separately within each group.
}
\section{Parameters}{

The parameters are inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{lower} :: \code{numeric(1)}\cr
Lower percentile bound (between 0 and 1). Default is \code{0.01}.
\item \code{upper} :: \code{numeric(1)}\cr
Upper percentile bound (between 0 and 1). Default is \code{0.99}.
\item \code{na.rm} :: \code{logical(1)}\cr
Whether to remove missing values when computing quantiles. Default is \code{TRUE}.
}
}

\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpWinsorize}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpWinsorize-new}{\code{PipeOpWinsorize$new()}}
\item \href{#method-PipeOpWinsorize-clone}{\code{PipeOpWinsorize$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpWinsorize-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpWinsorize-new}{}}}
\subsection{Method \code{new()}}{
Constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpWinsorize$new(id = "winsorize", param_vals = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"winsorize"}.}

\item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings. Default is \code{list()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpWinsorize-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpWinsorize-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpWinsorize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
