% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingHoldoutGapFix.R
\name{mlr_resamplings_holdout_gap_fix}
\alias{mlr_resamplings_holdout_gap_fix}
\alias{ResamplingHoldoutGapFix}
\title{Holdout with Gap (Fixed Split) Cross Validation}
\description{
Creates a single-split holdout resampling strategy where the training set consists of the first
\code{initial_window} observations (or groups), then a gap of \code{gap} observations (or groups) is skipped,
and the next \code{horizon} observations (or groups) form the test set.

If a group column is defined in the task (e.g., month), the splitting is done on the unique groups,
and the groups are mapped back to the row indices.
}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingHoldoutGapFix}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{\code{integer(1)}\cr
Returns the number of resampling iterations (always 1 for a holdout).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingHoldoutGapFix-new}{\code{ResamplingHoldoutGapFix$new()}}
\item \href{#method-ResamplingHoldoutGapFix-clone}{\code{ResamplingHoldoutGapFix$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingHoldoutGapFix-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingHoldoutGapFix-new}{}}}
\subsection{Method \code{new()}}{
Create a holdout with gap (fixed split) resampling instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingHoldoutGapFix$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingHoldoutGapFix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingHoldoutGapFix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingHoldoutGapFix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
