% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingHoldoutGapRatio.R
\name{mlr_resamplings_holdout_gap_ratio}
\alias{mlr_resamplings_holdout_gap_ratio}
\alias{ResamplingHoldoutGapRatio}
\title{Holdout with Ratio and Optional Gap Cross Validation}
\description{
Creates a single-split holdout resampling strategy where the training set is defined by a ratio of the data,
an optional gap is skipped after the training set, and the remaining observations form the test set.

If a group column is defined in the task (e.g. month), then the ratio and gap are applied on the unique groups,
and the groups are subsequently mapped back to the row indices.
}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingHoldoutGapRatio}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{\code{integer(1)}\cr
Returns the number of resampling iterations (always 1 for holdout).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingHoldoutGapRatio-new}{\code{ResamplingHoldoutGapRatio$new()}}
\item \href{#method-ResamplingHoldoutGapRatio-clone}{\code{ResamplingHoldoutGapRatio$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingHoldoutGapRatio-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingHoldoutGapRatio-new}{}}}
\subsection{Method \code{new()}}{
Create a holdout with gap resampling instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingHoldoutGapRatio$new(id = "holdout_gap_ratio")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character(1)}\cr
Identifier for the resampling strategy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingHoldoutGapRatio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingHoldoutGapRatio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingHoldoutGapRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
