% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingGapCV.R
\name{mlr_resamplings_gap_cv}
\alias{mlr_resamplings_gap_cv}
\alias{ResamplingGapCV}
\title{Rolling and Expanding Gap Cross Validation}
\description{
Creates a rolling or expanding gap cross-validation resampling strategy.
This resampling strategy is designed for time series data, where the training set is defined by an initial window,
a gap, and a horizon. The gap is the number of observations skipped between the training and test sets.
The horizon is the number of observations in the test set.
If a group column is defined in the task (e.g. month), then the initial window, gap, and horizon are applied on the unique groups,
and the groups are subsequently mapped back to the row indices.
}
\section{Parameters}{

The parameters are:
\itemize{
\item \code{initial_window} :: \code{integer(1)}\cr
The size of the initial training window. Default is \code{10}.
\item \code{horizon} :: \code{integer(1)}\cr
The size of the test set (horizon). Default is \code{5}.
\item \code{gap} :: \code{integer(1)}\cr
The number of observations to skip between the training and test sets. Default is \code{0}.
\item \code{step} :: \code{integer(1)}\cr
The number of observations to slide the training window forward for each fold. Default is \code{1}.
\item \code{rolling} :: \code{logical(1)}\cr
Whether to use a rolling (fixed-size) window or an expanding window. Default is \code{FALSE}.
}
}

\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingGapCV}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{\code{integer(1)}\cr
Returns the number of resampling iterations, depending on the
values stored in the \code{param_set}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingGapCV-new}{\code{ResamplingGapCV$new()}}
\item \href{#method-ResamplingGapCV-clone}{\code{ResamplingGapCV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingGapCV-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingGapCV-new}{}}}
\subsection{Method \code{new()}}{
Create an gap CV instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingGapCV$new(id = "gap_cv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character(1)}\cr
Identifier for the resampling strategy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingGapCV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingGapCV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingGapCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
